% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{baseline_terms}
\alias{baseline_terms}
\title{Extract baseline terms}
\usage{
baseline_terms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Extract baseline terms
}
