% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R, R/design_generics.R, R/event_model.R
\name{columns.Scale}
\alias{columns.Scale}
\alias{columns.ScaleWithin}
\alias{columns.RobustScale}
\alias{columns}
\alias{columns.event_model}
\title{Extract columns}
\usage{
\method{columns}{Scale}(x, ...)

\method{columns}{ScaleWithin}(x, ...)

\method{columns}{RobustScale}(x, ...)

columns(x, ...)

\method{columns}{event_model}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
Character vector of column names produced by the object.
}
\description{
Extract columns
}
\examples{
bs_basis <- BSpline(seq(0, 1, length.out = 5), degree = 3)
columns(bs_basis)
}
