% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model.R, R/covariate.R,
%   R/design_generics.R, R/feature.R
\name{construct.baselinespec}
\alias{construct.baselinespec}
\alias{construct.covariatespec}
\alias{construct}
\alias{construct.featurespec}
\title{Construct method}
\usage{
\method{construct}{baselinespec}(x, model_spec, ...)

\method{construct}{covariatespec}(x, model_spec, sampling_frame = NULL, ...)

construct(x, ...)

\method{construct}{featurespec}(x, model_spec, ...)
}
\arguments{
\item{x}{The object.}

\item{model_spec}{A model specification object (used by some methods).
For baselinespec: typically a sampling_frame or list containing one.
For hrfspec/covariatespec: contains data and other model information.}

\item{...}{Additional arguments.}

\item{sampling_frame}{A sampling_frame object (used by covariatespec method).}
}
\value{
A constructed object; return type depends on method.
}
\description{
Construct method
}
\examples{
sframe <- fmrihrf::sampling_frame(blocklens = 5, TR = 1)
drift_spec <- baseline(degree = 2, basis = "poly")
construct(drift_spec, sframe)
}
