% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/feature.R
\name{contrasts}
\alias{contrasts}
\alias{contrasts.feature_term}
\title{Extract contrasts}
\usage{
contrasts(x, ...)

\method{contrasts}{feature_term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of contrast specifications.
}
\description{
Extract contrasts
}
\examples{
des <- data.frame(
  onset = c(0, 10, 20, 30),
  run = 1,
  cond = factor(c("A", "B", "A", "B"))
)
sframe <- fmrihrf::sampling_frame(blocklens = 40, TR = 1)
cset <- contrast_set(
  diff = column_contrast(pattern_A = "cond.A", pattern_B = "cond.B", name = "diff")
)
emod <- event_model(onset ~ hrf(cond, contrasts = cset),
                    data = des, block = ~run, sampling_frame = sframe)
contrasts(emod)
}
