% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{design_map}
\alias{design_map}
\title{Compute design map}
\usage{
design_map(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot2 object visualizing the design matrix.
}
\description{
Compute design map
}
