% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_map.R
\name{.correlation_map_common}
\alias{.correlation_map_common}
\title{Design Matrix Visualization Methods}
\usage{
.correlation_map_common(
  DM,
  method = c("pearson", "spearman"),
  half_matrix = FALSE,
  absolute_limits = TRUE,
  ...
)
}
\value{
A ggplot2 heatmap of the correlation matrix.
}
\description{
This file contains methods for visualizing design matrices as heatmaps
using ggplot2.
}
\keyword{internal}
