% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrihrf-imports.R
\name{fmrihrf-reexports}
\alias{fmrihrf-reexports}
\title{fmrihrf reexports}
\value{
See the corresponding fmrihrf function documentation.
}
\description{
Re-exported functions from the fmrihrf package for convenience.
See the upstream fmrihrf documentation for details on usage and return
values of each function.
}
