% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event-classes.R
\name{is_continuous}
\alias{is_continuous}
\alias{is_continuous.event}
\title{Check if continuous}
\usage{
is_continuous(x, ...)

\method{is_continuous}{event}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Check if continuous
}
