% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_registry.R
\name{is_external_hrfspec}
\alias{is_external_hrfspec}
\title{Check if a Class is a Registered External HRF Specification}
\usage{
is_external_hrfspec(x)
}
\arguments{
\item{x}{An object or character string class name}
}
\value{
Logical indicating if the class is registered as an external HRF spec
}
\description{
Check if a Class is a Registered External HRF Specification
}
