% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_registry.R
\name{list_external_hrfspecs}
\alias{list_external_hrfspecs}
\title{List All Registered External HRF Specifications}
\usage{
list_external_hrfspecs()
}
\value{
A character vector of registered class names
}
\description{
List All Registered External HRF Specifications
}
