% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_registry.R
\name{list_registered_bases}
\alias{list_registered_bases}
\title{List Registered Parametric Basis Classes}
\usage{
list_registered_bases()
}
\value{
Character vector of registered class names.
}
\description{
List Registered Parametric Basis Classes
}
\examples{
list_registered_bases()
}
