% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{predict.ParametricBasis}
\alias{predict.ParametricBasis}
\alias{predict.Standardized}
\alias{predict.Poly}
\alias{predict.BSpline}
\alias{predict.Ident}
\alias{predict.Scale}
\alias{predict.ScaleWithin}
\alias{predict.RobustScale}
\title{Predict from a ParametricBasis object}
\usage{
\method{predict}{ParametricBasis}(object, newdata, ...)

\method{predict}{Standardized}(object, newdata, ...)

\method{predict}{Poly}(object, newdata, ...)

\method{predict}{BSpline}(object, newdata, ...)

\method{predict}{Ident}(object, newdata, ...)

\method{predict}{Scale}(object, newdata, ...)

\method{predict}{ScaleWithin}(object, newdata, newgroup, ...)

\method{predict}{RobustScale}(object, newdata, ...)
}
\arguments{
\item{object}{ParametricBasis object.}

\item{newdata}{Numeric vector to transform.}

\item{...}{Additional arguments.}

\item{newgroup}{Optional factor for group-dependent bases.}
}
\value{
A numeric matrix with transformed values (one column per basis component).
}
\description{
Dispatch to the appropriate method for transforming new data
according to a specific parametric basis.
}
