% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrihrf-imports.R, R/fmrihrf-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{HRF}
\alias{as_hrf}
\alias{gen_hrf}
\alias{sampling_frame}
\alias{hrf_spmg1}
\alias{regressor}
\alias{feature_regressor}
\alias{blocklens}
\alias{samples}
\alias{global_onsets}
\alias{evaluate}
\alias{onsets}
\alias{durations}
\alias{blockids}
\alias{nbasis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fmrihrf}{\code{\link[fmrihrf]{as_hrf}}, \code{\link[fmrihrf]{blockids}}, \code{\link[fmrihrf]{blocklens}}, \code{\link[fmrihrf]{durations}}, \code{\link[fmrihrf]{evaluate}}, \code{\link[fmrihrf]{feature_regressor}}, \code{\link[fmrihrf]{gen_hrf}}, \code{\link[fmrihrf]{global_onsets}}, \code{\link[fmrihrf:HRF-class]{HRF}}, \code{\link[fmrihrf]{hrf_spmg1}}, \code{\link[fmrihrf]{nbasis}}, \code{\link[fmrihrf]{onsets}}, \code{\link[fmrihrf]{regressor}}, \code{\link[fmrihrf]{samples}}, \code{\link[fmrihrf]{sampling_frame}}}
}}

