% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{ScaleWithin}
\alias{ScaleWithin}
\title{Z-score within groups}
\usage{
ScaleWithin(x, g)
}
\arguments{
\item{x}{numeric vector}

\item{g}{grouping factor / character / integer of same length as x}
}
\value{
An object of class \code{ScaleWithin} (a \code{ParametricBasis}).
}
\description{
Z-score within groups
}
