% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{term_indices}
\alias{term_indices}
\title{Extract term indices}
\usage{
term_indices(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Extract term indices
}
