% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{term_matrices}
\alias{term_matrices}
\title{Extract term matrices}
\usage{
term_matrices(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A list of matrices/tibbles, one per term.
}
\description{
Extract term matrices
}
