% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R, R/design_generics.R
\name{columns.Scale}
\alias{columns.Scale}
\alias{columns.ScaleWithin}
\alias{columns.RobustScale}
\alias{columns}
\title{Extract columns}
\usage{
\method{columns}{Scale}(x, ...)

\method{columns}{ScaleWithin}(x, ...)

\method{columns}{RobustScale}(x, ...)

columns(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
Character vector of column names produced by the object.
}
\description{
Extract columns
}
