% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_vector.R
\name{conditions}
\alias{conditions}
\alias{conditions.event_term}
\title{Extract conditions from a design object}
\usage{
conditions(x, drop.empty = TRUE, expand_basis = FALSE, ...)

\method{conditions}{event_term}(x, drop.empty = TRUE, expand_basis = FALSE, ...)
}
\arguments{
\item{x}{The object to extract conditions from.}

\item{drop.empty}{Logical whether to drop conditions with no events (default: TRUE).}

\item{expand_basis}{Logical whether to expand basis functions (default: FALSE).}

\item{...}{Additional arguments.}
}
\description{
Extract conditions from a design object
}
