% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{construct}
\alias{construct}
\title{Construct method}
\usage{
construct(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A constructed object; return type depends on method.
}
\description{
Construct method
}
