% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_model.R
\name{contrast_weights}
\alias{contrast_weights}
\alias{contrast_weights.convolved_term}
\alias{contrast_weights.event_model}
\title{Compute contrast weights}
\usage{
contrast_weights(x, ...)

\method{contrast_weights}{convolved_term}(x, ...)

\method{contrast_weights}{event_model}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of contrast weight objects or matrices.
}
\description{
Compute contrast weights
}
\details{
If the weight matrices returned by a contrast specification contain row
names, these are matched to the column names of the corresponding term in
the design matrix.  This allows contrasts to target only a subset of term
levels.
}
