% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{contrasts}
\alias{contrasts}
\title{Extract contrasts}
\usage{
contrasts(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of contrast specifications.
}
\description{
Extract contrasts
}
