% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_vector.R
\name{convolve}
\alias{convolve}
\alias{convolve.event_term}
\title{Convolve events with a hemodynamic response function}
\usage{
convolve(
  x,
  hrf,
  sampling_frame,
  drop.empty = TRUE,
  summate = TRUE,
  precision = 0.1,
  ...
)

\method{convolve}{event_term}(
  x,
  hrf,
  sampling_frame,
  drop.empty = TRUE,
  summate = TRUE,
  precision = 0.3,
  ...
)
}
\arguments{
\item{x}{The events to convolve.}

\item{hrf}{The hemodynamic response function.}

\item{sampling_frame}{The sampling frame.}

\item{drop.empty}{Logical indicating whether to drop columns with all zeros.}

\item{summate}{Logical indicating whether to sum convolved signals.}

\item{precision}{Numeric specifying the temporal precision for convolution.}

\item{...}{Additional arguments.}
}
\value{
A matrix-like (often tibble) of convolved regressors.
}
\description{
Convolve events with a hemodynamic response function
}
