% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{correlation_map}
\alias{correlation_map}
\title{Compute correlation map}
\usage{
correlation_map(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot2 object visualizing regressor correlations.
}
\description{
Compute correlation map
}
