% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_registry.R
\docType{data}
\name{.fmridesign_extensions}
\alias{.fmridesign_extensions}
\title{Extension Registry for External HRF Specifications}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.fmridesign_extensions
}
\description{
This file provides an extension mechanism for packages to register
their own HRF specification types with fmridesign.
}
\keyword{datasets}
