% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/feature.R
\name{event_table}
\alias{event_table}
\alias{event_table.feature_term}
\title{Extract event table}
\usage{
event_table(x, ...)

\method{event_table}{feature_term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A data.frame/tibble of event rows.
}
\description{
Extract event table
}
\examples{
term <- event_term(
  list(condition = factor(c("A", "B", "A"))),
  onsets = c(0, 10, 20),
  blockids = c(1, 1, 1)
)
event_table(term)
}
