% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{event_table}
\alias{event_table}
\title{Extract event table}
\usage{
event_table(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\description{
Extract event table
}
