% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{event_terms}
\alias{event_terms}
\title{Extract event terms}
\usage{
event_terms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of event term objects.
}
\description{
Extract event terms
}
