% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_model.R, R/feature.R
\name{Fcontrasts}
\alias{Fcontrasts}
\alias{Fcontrasts.convolved_term}
\alias{Fcontrasts.event_model}
\alias{Fcontrasts.feature_term}
\title{Compute F-contrasts}
\usage{
Fcontrasts(x, ...)

\method{Fcontrasts}{convolved_term}(x, ...)

\method{Fcontrasts}{event_model}(x, ...)

\method{Fcontrasts}{feature_term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of matrices with F-contrast weights.
}
\description{
Compute F-contrasts
}
\details{
Row names of the contrast matrices can specify which levels of the term are
tested.  Any matching is done against the design matrix column names.
}
\examples{
des <- data.frame(
  onset = c(0, 10, 20, 30),
  run = 1,
  cond = factor(c("A", "B", "A", "B"))
)
sframe <- fmrihrf::sampling_frame(blocklens = 40, TR = 1)
emod <- event_model(onset ~ hrf(cond), data = des, block = ~run, sampling_frame = sframe)
names(Fcontrasts(emod))
}
