% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_model.R
\name{Fcontrasts}
\alias{Fcontrasts}
\alias{Fcontrasts.convolved_term}
\alias{Fcontrasts.event_model}
\title{Compute F-contrasts}
\usage{
Fcontrasts(x, ...)

\method{Fcontrasts}{convolved_term}(x, ...)

\method{Fcontrasts}{event_model}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A named list of matrices with F-contrast weights.
}
\description{
Compute F-contrasts
}
\details{
Row names of the contrast matrices can specify which levels of the term are
tested.  Any matching is done against the design matrix column names.
}
