% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrihrf-reexports.R
\name{fmrihrf-reexports}
\alias{fmrihrf-reexports}
\alias{onsets}
\alias{durations}
\alias{blockids}
\title{Functions re-exported from fmrihrf}
\usage{
onsets(x, ...)

durations(x, ...)

blockids(x, ...)
}
\description{
These functions are re-exported from the fmrihrf package.
Note: When both packages are loaded, R will show masking warnings.
This is expected and harmless - the functions work identically.
}
\details{
The following generics are re-exported:
\itemize{
\item \code{\link[fmrihrf]{onsets}}: Extract onset times from event objects
\item \code{\link[fmrihrf]{durations}}: Extract durations from event objects
\item \code{\link[fmrihrf]{blockids}}: Extract block identifiers
}

fmridesign adds S3 methods for these generics to work with:
\itemize{
\item event_term objects
\item convolved_term objects
\item event_model objects
}
}
\keyword{internal}
