% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event-classes.R
\name{is_categorical}
\alias{is_categorical}
\alias{is_categorical.event}
\title{Check if categorical}
\usage{
is_categorical(x, ...)

\method{is_categorical}{event}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
Logical scalar indicating whether \code{x} is categorical.
}
\description{
Check if categorical
}
