% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{longnames}
\alias{longnames}
\title{Extract longnames}
\usage{
longnames(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
Character vector of long (fully qualified) names.
}
\description{
Extract longnames
}
