% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-formula.R
\name{nbasis.hrfspec}
\alias{nbasis.hrfspec}
\title{Get number of basis functions from hrfspec}
\usage{
\method{nbasis}{hrfspec}(x, ...)
}
\arguments{
\item{x}{An hrfspec object}

\item{...}{Additional arguments (unused)}
}
\value{
The number of basis functions
}
\description{
Get number of basis functions from hrfspec
}
