% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_map.R
\name{plot.event_model}
\alias{plot.event_model}
\title{Plot Event Model}
\usage{
\method{plot}{event_model}(x, term_name = NULL, ...)
}
\arguments{
\item{x}{An \code{event_model} object.}

\item{term_name}{Character. Name of specific term to plot. If NULL, plots all terms.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A ggplot2 object showing the predicted BOLD timecourses.
}
\description{
Creates a line plot visualization of the predicted BOLD response for each
regressor in an event_model object.
}
