% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_vector.R
\name{regressors}
\alias{regressors}
\alias{regressors.event_term}
\title{Extract regressors}
\usage{
regressors(x, ...)

\method{regressors}{event_term}(x, hrf, sampling_frame, summate = FALSE, drop.empty = TRUE, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}

\item{hrf}{HRF function}

\item{sampling_frame}{sampling_frame object}

\item{summate}{Logical; sum HRF responses}

\item{drop.empty}{Logical; drop empty conditions}
}
\value{
Character vector of regressor names for \code{x}.
}
\description{
Convolve the event-term design matrix with an HRF and return the
resulting regressors.
}
