% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{shortnames}
\alias{shortnames}
\title{Extract shortnames}
\usage{
shortnames(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
Character vector of short names.
}
\description{
Extract shortnames
}
