% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{split_by_block}
\alias{split_by_block}
\title{Split by block}
\usage{
split_by_block(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A list split by block/run.
}
\description{
Split by block
}
\examples{
des <- data.frame(
  onset = c(0, 10, 20, 30, 5, 15, 25, 35),
  run = c(1, 1, 1, 1, 2, 2, 2, 2),
  cond = factor(c("A", "B", "A", "B", "A", "B", "A", "B"))
)
sframe <- fmrihrf::sampling_frame(blocklens = c(40, 40), TR = 1)
emod <- event_model(onset ~ hrf(cond), data = des, block = ~run, sampling_frame = sframe)
block_list <- split_by_block(emod)
length(block_list)
}
