% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R
\name{split_by_block}
\alias{split_by_block}
\title{Split by block}
\usage{
split_by_block(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments.}
}
\value{
A list split by block/run.
}
\description{
Split by block
}
