% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_generics.R, R/event_vector.R
\name{split_onsets}
\alias{split_onsets}
\alias{split_onsets.event_term}
\title{Split onsets}
\usage{
split_onsets(x, sframe, global = FALSE, blocksplit = FALSE, ...)

\method{split_onsets}{event_term}(x, sframe, global = FALSE, blocksplit = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{sframe}{The sampling frame object containing timing information.}

\item{global}{Whether onsets are in global time units (across all runs).}

\item{blocksplit}{Whether to split onsets by blocks.}

\item{...}{Additional arguments.}
}
\description{
Split onsets
}
